+++ src/main/kotlin/com/nullpointercats/sys/adopta/user/dto/response/LoginResponse.kt
@@ -1,4 +1,4 @@
-package com.nullpointercats.sys.adopta.dto.response
+package com.nullpointercats.sys.adopta.user.dto.response
 
 /**
  * DTO used to send authentication information back to the client
