+++ src/main/kotlin/com/nullpointercats/sys/adopta/user/dto/response/RegisterResponse.kt
@@ -1,4 +1,4 @@
-package com.nullpointercats.sys.adopta.dto.response
+package com.nullpointercats.sys.adopta.user.dto.response
 
 /**
  * DTO used to send user registration confirmation
